Refactor Git::config
The Git module has config and global_config methods that are available
only when Git module is included. Refactored the code to remove
duplication and to make it more explicit we are managing `git config`
not the library setup.
